tasks.py: finish the tutorial
The tutorial includes login in as Maria, downloading the spreadsheet and filling each row of the spreadsheet
into the form. Then taking a screenshot of that page and finally generating a PDF with all names filled out.

This is interesting to know that there are libraries capable of doing that.

Signed-off-by: Charles Oliveira <18oliveira.charles@gmail.com>